Enough unit tests to complete coverage of SmallMultiplesHelper
